Refactor query builder and tests
