+++ gqlalchemy/__init__.py
@@ -12,6 +12,8 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+from .query_builders import neo4j_query_builder, query_builder  # noqa F401
+
 from .disk_storage import SQLitePropertyDatabase  # noqa F401
 from .exceptions import GQLAlchemyWarning, GQLAlchemyError  # noqa F401
 from .instance_runner import (  # noqa F401
@@ -37,7 +39,7 @@ from .models import (  # noqa F401
     Field,
 )
 from .vendors.neo4j import Neo4j  # noqa F401
-from .query_builder import (  # noqa F401
+from .query_builders.query_builder import (  # noqa F401
     Call,
     Create,
     InvalidMatchChainException,
@@ -50,7 +52,7 @@ from .query_builder import (  # noqa F401
     Unwind,
     With,
 )
-from .neo4j_query_builder import Neo4jQueryBuilder  # noqa F401
+from .query_builders.neo4j_query_builder import Neo4jQueryBuilder  # noqa F401
 
 from .exceptions import GQLAlchemyWarning, GQLAlchemyError  # noqa F401
 from pydantic import validator  # noqa F401
