+++ tests/integration/test_graph_algorithms.py
@@ -13,7 +13,7 @@
 # limitations under the License.
 
 from gqlalchemy.graph_algorithms.query_modules import QueryModule, parse_query_module_signature
-from gqlalchemy.query_builder import QueryBuilder
+from gqlalchemy.query_builders.query_builder import QueryBuilder
 from gqlalchemy import Memgraph
 
 from typing import List
